Decoupled ColumnGroup logic from ColumnState
